+++ flowforge-spring-boot-autoconfigure/src/main/java/io/flowforge/spring/annotations/FlowWorkflow.java
@@ -2,7 +2,7 @@ package io.flowforge.spring.annotations;
 
 import java.lang.annotation.*;
 
-@Target(ElementType.METHOD)
+@Target({ElementType.METHOD, ElementType.TYPE})
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
 public @interface FlowWorkflow {