+++ flowforge-spring-boot-autoconfigure/src/main/java/io/flowforge/spring/workflow/WorkflowDefinition.java
@@ -1,0 +1,12 @@
+package io.flowforge.spring.workflow;
+
+import io.flowforge.spring.dsl.FlowDsl;
+import io.flowforge.workflow.plan.WorkflowExecutionPlan;
+
+/**
+ * Class-based workflow definition contract.
+ */
+public interface WorkflowDefinition {
+
+    WorkflowExecutionPlan define(FlowDsl dsl);
+}